#23 se termino de realizar las politicas

+++ app/Http/Controllers/ArticleController.php
@@ -19,10 +19,12 @@ class ArticleController extends Controller
 
     public function index()
     {
+        #$this->authorize('viewAny', Article::class);
         return new ArticleCollection(Article::paginate(20));
     }
     public function show(Article $article)
     {
+        $this->authorize('view', $article);
         return response()->json(new ArticleResource($article),200);
     }
 
@@ -34,6 +36,7 @@ class ArticleController extends Controller
     public function store(Request $request)
     {
 
+        $this->authorize('create', Article::class);
         $request->validate([
             'title' => 'required|string|unique:articles|max:255',
             'body' => 'required',
@@ -52,6 +55,8 @@ class ArticleController extends Controller
     }
     public function update(Request $request,  Article $article)
     {
+        $this->authorize('update', $article);
+
         $request->validate([
             'title' => 'required|string|unique:articles,title,' . $article->id . '|max:255',
             'body' => 'required',
@@ -63,8 +68,10 @@ class ArticleController extends Controller
     }
     public function delete(Article $article)
     {
-        $article->delete();
 
+        $this->authorize('delete', $article);
+
+        $article->delete();
         return response()->json(null, 204);
     }
 }
